Made the  "zoom to selection" much faster. It was previously disabling the zoom selected button when no selection was active. This forced some strange layout manager issue involving sorting all components for determining the focus ordering. This could take a long time.
Fixed a problem with the new user supplied data and memory usage reading when start time is after the begin event, but before the
new type of record for supplied data or memory.

+++ projections/analysis/LogEntry.java
@@ -81,12 +81,12 @@ class LogEntry
 	return log;
     }
 
-	public int userSuppliedValue() {
-	  return userSupplied.intValue();
+	public Integer userSuppliedValue() {
+	  return userSupplied;
 	}
 	
-	public int memoryUsage() {
-	  return memoryUsage.intValue();
+	public Integer memoryUsage() {
+	  return memoryUsage;
 	}
 
 	public void setUserSupplied(int userSuppliedValue) {
